+++ product-service/handler.ts
@@ -1,4 +1,4 @@
-import { JewelleryProductService } from '@libs/services/productsService';
+import { JewelleryProductService } from '@libs/services/jewelleryProductService';
 import * as handlers from './src/functions';
 
 const productService = new JewelleryProductService();
