+++ product-service/src/libs/services/jewelleryProductService/index.ts
@@ -1,0 +1,1 @@
+export { default as JewelleryProductService } from './jewelleryProductService';
