+++ product-service/src/libs/services/jewelleryProductService/products-mock.ts
