+++ product-service/src/libs/services/productsService/productService.ts
